fix(Money.java): add Exception handling for over MAX_INT moneyInput MAX_INT를 초과한 값에서 에러 메시지가 제대로 출력되지 않는 오류를 해결했습니다.

+++ src/main/java/lotto/model/Money.java
@@ -20,7 +20,7 @@ public class Money {
         validateEmptyInput(moneyInput);
         validateMoneyLength(moneyInput);
         validateMoneyNumeric(moneyInput);
-        validateMoneyInRange(Integer.parseInt(moneyInput));
+        validateMoneyInRange(Long.parseLong(moneyInput));
         validateMoneyDividedTByThousand(Integer.parseInt(moneyInput));
         return Integer.parseInt(moneyInput);
     }
@@ -37,7 +37,7 @@ public class Money {
         }
     }
 
-    void validateMoneyInRange(int money) {
+    void validateMoneyInRange(long money) {
         if (money < MIN_MONEY_RANGE || money > MAX_MONEY_RANGE) {
             Exception.raiseInvalidMoneyRangeException();
         }
